Somewhat working parameter chooser implementation
Getting the UI layout to behave was a nightmare; tool buttons
in the first column just wouldn't take the right size, or rather, only
when a group was expanded (!).

In the end, the issue appeared to be that the column size is calculated
without considering the QTreeView.indentation (the extra space for the
tree expanding triangles). I couldn't find a fix that works with
widget items, so in the end I gave up and put the buttons to the right.