parameters: Rewrite parameter change logic without function pointers
It is not clear what exactly was going on, but it seems like
the function pointers played some role in a kernel CPU memory
corruption issue.

Now that we have resigned ourselves to duplicating the store/handle
classes for every data type anyway, just storing references to the
handles directly isn't an issue.