plots: Make single-point scans robust
Previously, single-point (Single/Repeat) plots relied on not being
launched in the middle of a point update, and the first (only) point
could be missed for Single scans. This wasn't an issue for us so far,
as kernel compilation times usually trumped applet startup times.

This reverts 60d205bbf549f7e9a7dd26a3664132b432d81b87 as well, i.e.
adds back point_phase, as otherwise there is no guarantee about the
order in which individual points are pushed. (This could be avoided
in the future by buffering the push() results instead, and only
emitting the datasets in a known order once the point is finished.)