dashboard.argument_editor: Error on unfinished dataset sync without crashing
Previously, if defaults were evaluated before the dashboard had finished
syncing datasets from master (for instance because the user pressed Recompute
Arguments immediately after starting the dashboard with some 100s of MiB of
master datasets), the dashboard would crash entirely.

This was already converted into an error message for the case where
experiments are opened previously, but not for resetting entries to the
default value.