experiment: Attach identity to ParamHandles
Each ParamHandle instance always corresponded to exactly one
attribute in one owning fragment. Previously, this relationship
was only discoverable from the fragment side (and only internally);
user-facing code needed to accept a reference to the owning
fragment and a parameter name.

The motivation behind the original choice was mainly to clearly
delineate which dependencies should preferentially be used in the
design by making it syntactically hard/impossible to go against that
(e.g. conceptually a handle shouldn't be used without "owning" a
reference to the parent fragment). It does make APIs more cumbersome
to use, however, so this commit adds the information about this
mapping also to the fragments, in preparation of functionality
make use of this.

Existing APIs (rebinding, subscans, …) should probably be modified
to support using handles directly as well.