experiment: Properly implement/test Subscan.run(execute_default_analyses=True)
GitHub: Fixes #211. This should finally be a fairly clean
solution for the currently badly supported use case where one
of a few different possible subscans is selected only after
build_fragment() – result channels are not created for
analysis results if their axes are only a subset of those
available, but if at runtime only that subset is scanned, those
results are still retuned from Subscan.run().