LICENSE: GPLv3 -> LGPLv3
I had missed that ARTIQ actually uses the LGPL, which would have
made contributing code back into ARTIQ harder than necessary. The
setup.py metadata had already been listing the license as LGPL for
a long time, and thus far I am the only contributor of changes that
are large enough to be copyrightable, so it's an easy and
uncontroversial fix to make.