Allow `dataset("…")` without fallback as default value
This will cause experiments to fail to be built for
datasets that don't exist, and the dashboard argument
editors to potentially fail to open if the master dataset
sync has not yet completed. For those reasons, a default
value was originally required, but several users prefer
the "louder" failure behaviour over the warning + fallback.

GitHub: Fixes #209.