experiment: Don't execute {device,host}_{setup,cleanup}() for subscanned fragments
Previously, an empty version of the methods would need to be provided
manually in the parent fragment to avoid the scanned fragment to be
automatically set up/torn down as part of the default recursive
implementation. This is especially annoying as it might cause a lot
of extraneous kernel compilations.

GitHub: Fixes #288.