dataset_janitor: Make sure to explicitly close all connections on failure
ABaQuS has been sporadically seeing "Too many open files" on the
master process, likely because of the RPC connection. It is still
unclear why the RPC calls failed for them in the first place, though;
in lab one, we do not observe the failures that triggered the leak
on the recovery path.