ARTIQ 8 (Git master) compatibility
The SimpleApplet API was changed in m-labs/artiq@fe0f6d8a2c0
(with the new dataset metadata being introduced) in a
backwards-incompatible fashion.

There is now also a built-in way to call set_dataset() without
a separate RPC connection. This also turned out to be
backwards-incompatible, as the --port options would conflict.

Currently, this is not backwards-compatible to ARTIQ 7.