plots.xy_1d: fixup annotations
As we want stable series colours when hiding/showing series from the context menu,
the series_idx is not monotonic. To maintain matching colours of the annotations,
this had to be added to _XYSeries.
Moreover, 'location' annotations are now shown once on each pane, unless associated_channels
are specified. In that case, the annotation is only added to the pane(s) where these
channel(s) are shown. To make this information available, the pane_idx had to be added to
_XYSeries.