Revise EnumParam implementation on experiment side, with defaults in mind
Needed a clear convention what qualifies as a default value. Either given as Enum object
or as string that evaluates to the name of an Enum member. This allows the value of the
Enum members to be arbitrary objects.
For displaying purposes, the spec includes a dictionary mapping Enum member names to
strings. If the values happen to be (descriptive) strings, those will be shown in the
dashboard and applets. Otherwise, the Enum member names are mapped to themselves.