experiment.parameters: Avoid type hint forward reference for linters [nfc]
This is technically a circular reference, so (in a pre-PEP563 world) the
type[<string>] syntax was used. With star imports gone from the file,
flake8 catches that it is not actually defined, though. Importing the
fragment module would cause circular dependency issues during module
loading, though, so just drop the annotation for now.