experiment.scan_runner: Fix on-kernel scan compilation after 577b3fb
The issue is that ARTIQ does not convert built-in types to
artiq.compiler types when used as arguments to compiler types,
which was the case here for TTuple/TList.

We desperately need kernel tests as part of the test suite; this
would have been caught by the most straightforward scan execution
test.