Always log the adjacent source for rebinds rather than the ultimate resolved source.
This is not sensitive to the order in which
rebinds happen, and the ultimate source can be
found using _find_param_source. This doesn't
currently matter, but future code might find this
behaviour less surprising.

Note that _rebound_subfragment_params *is*
sensitive to the order in which binds happen.
This does not create redundant stores however, because init_params only
reads _get_all_handles_for_param for free params.