plots: Unify subscans and alternate plots into dockable panes
Opening separate windows for subscans was never a good UI choice, it
just was the path of least resistance at the time. Now, subscans are
displayed right next to the main scan.

In the future, this could be combined with saving/restoring of the
entire geometry between runs of the same scan to avoid having to
reopen the subscans every time.

Multiple nested subscans are not yet tested.

This does not address other subscan UI issues (e.g. having to re-select
points to get the plots to show up).