plots.xy_1d: Improved point highlighting; slightly decreased point size
Highlight borders are now shown for all series and are persistent as new
points come in; they are drawn on top of other points. Also fixed the
selection to target the closest point (noticeable with very dense scans).

The point size is a small tweak; now that the default is for every
channel to be in its own pane, each pane tends to be smaller.