plots: Immediately check whether model already has schema/data
The previous design was an attempt to streamline the implementation by
having the models guarantee to emit a changed/… signal later (once
control was returned to the event loop, i.e. plots had a chance to
connect their slots), even if the model was already complete at
construction time.

This turned out to add more complexity than it removed, and did not work
properly for subscans (causing the user to have to switch points to
actually have the plot load).

After this, the plots themselves defer the initialisation until later,
as the calling code does not expect ready() to be immediately signalled.
This could be similarly simplified in the future.