plots: Make sure subscan plot objects are deleted after closing
I'd have liked to avoid the explicit deleteLater() call, but the
subplots were not properly cleaned up.

I am committing this long after making this change, and my notes on
this are not as good as I would like. To test this in the future,
add debug prints to check how many live model subscribers exist
(e.g. in the update…() slots), and make sure they do not accumulate
when closing/reopening subscans and toggling between top-level scan
points.