plots: Show context menu also without panes
This makes subplot docks accessible even if the top-level plot does not
have any non-opaque channels. Having access to the context menu is
enough for this to be functional for rolling plots, but we'll need to
come up with a way to do point selection for "empty" xy_1d plots.