argument_editor: Ignore apply_color() for now until API stabilises
On latest m-labs/master, apply_color() was changed to take only one
argument. This is a hotfix to allow ndscan to be used with current
ARTIQ master, and also hedges against future API changes before the
next release as Windows 11 style compatibility is figured out.
(Colours won't be applied to the ndscan widgets, but this is not
essential for operation.)