Update UI and scan generators for centred scans
Adds a new scan generator: CentreRefiningGenerator. This acts like the existing RefiningGenerator, only with the scan specified by centre and half span.
Changes `Refiningenerator` to silently drop points outside of the limits, such that the centre of scans doesn't change if a limit is hit.
Updates UI so that the centre-span options now emit either "centre_refining" or "centre_span" generators