argument_editor: Remove superfluous attempt_read_from_axis() type checking
A matching type is already ensured by the parameter schema, so don't
need to check for that. A change in parameter schemas might lead to
unexpected behaviour, but that's fine (same behaviour with
Recompute All Arguments being required in general).

Rather, when we support list scans for bools/enums (no reason why that
shouldn't be), we'll need to check the _values_ to figure out whether
to show the current all-option scan or a customised list scan, but that
can wait until those list scan options are implemented.